retrieves domino reference
no claims of efficiency are being made. i'm not considering performance
at this stage.

- uses d.get() several times, expect extra lookups to be wasteful.
- uses Integer and int, haven't considered if null overhead exists.

in terms of coding conventions, i simply decided to put get() in the
Domino class to return this;  Am uncertain if there's a preferential way
to `@Override` the contains() method from extending list. I'll set that
aside for now also.